Some more fixes for d3d8 and tab control.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@963 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CGUITabControl.cpp
@@ -458,7 +458,7 @@ void CGUITabControl::removeChild(IGUIElement* child)
 	// reassign numbers
 	if (isTab)
 	{
-		for (i=0; i<(s32)Tabs.size(); ++i)
+		for (i=0; i<Tabs.size(); ++i)
 			if (Tabs[i])
 				Tabs[i]->setNumber(i);
 	}
